Move test files under `test`
